+++ qqq-sample-project/src/main/java/com/kingsrook/sampleapp/SampleCli.java
@@ -50,6 +50,7 @@ public class SampleCli
    {
       try
       {
+         SampleMetaDataProvider.primeTestDatabase("prime-test-database.sql");
          QInstance qInstance = new SampleMetaDataProvider().defineQInstance();
          return (run(qInstance, args));
       }